Move CommandWordEnum to logic package
CommandWordEnum doesn't really belong in the logic.commands package
since it contains references to objects from both the logic.commands and
logic.parser packages.

This change also has the side benefit of making the class diagram less
messy.
